250428 edits to preonic
