--- keyboards/preonic/keymaps/daemontcat/rules.mk
@@ -1,1 +1,0 @@
-SRC += muse.c
